arc: implement `pack` command
Implement archive creation/modification using a manifest (similar to
archive packing in alice-tools).

Example of a manifest:

    #ARCPACK --mod-arc=MES.ARC.IN
    MES.ARC
    START.MES
    MAIN.MES

This will cause the program to read `MES.ARC.IN` and substitute the
files `START.MES` and `MAIN.MES` in the archive with files from the
filesystem, and then output a new archive named `MES.ARC`. The files
will be automatically LZSS encoded because they have the `.MES`
extension.

Because different games use different encryption keys for the archive
index, the `arc pack` command accepts a `--game` argument to select the
appropriate keys. It also accepts a `--key` argument which takes a
hexadecimal representation of the keys (this string can be obtained by
running `arc extract --key` on a archive).

Also included in this commit are the `lzss compress` and
`lzss decompress` commands. These exist mainly for testing purposes.