text: use line statement instead of padding
Use the line statement to terminate lines instead of padding to force a
new line. This way the user does not have to use the `columns`
directive.

However, if the user does NOT provide a `columns` directive, the last
column cannot be used. This is because the engine automatically
introduces a newline when a character is written to the last column.

If the `columns` directive is used, the last column can still be used,
and a warning will be printed if a line exceeds the configured number of
columns.