anim: support decompiling/compiling S4 files
Add the `anim decompile` and `anim compile` commands, which operate on
S4 format animation files.

An S4 file consists of one or more bytecode streams which are executed
to render the animation. The bytecode format is relatively simple, and
uses the same graphics primitives that are present in the AI5 engine.

This implementation can decompile/compile every S4 file present in
YU-NO. However, there are a number of S4 files in that game which seem
to be broken. These files appear to contain multiple bytecode streams,
but upon closer inspection, each stream after the first points to a
location near the end of the first stream.

This occurs in files which use the INNER_LOOP_START instruction, and the
stream pointer is off by the number of occurrences of this instruction.
Probably this is an off-by-one error in elf's encoder.

To handle these broken files, we refuse to read any stream which
overlaps with a previous stream, treating it as empty instead.