anim: fix timing issue
It seems like draw calls are batched up, but every other instruction
starts a new frame? E.g. in S42.S4 from Dragon Knight 4:

    LOOP_START 13;
        COPY 1(112, 288) -> 0(352, 104) @ (104, 96);
        STALL 1;
        COPY 1(0, 288) -> 0(352, 104) @ (104, 96);
    LOOP_END;

We need to introduce a new frame when we hit LOOP_END, otherwise the
second COPY will never be seen. On the other hand, immediately following
this is:

    LOOP_START 10;
        COPY 1(112, 288) -> 0(352, 104) @ (104, 96);
        COPY 1(0, 288) -> 0(352, 104) @ (104, 96);
    LOOP_END;

Logically, we need a new frame between each COPY for the same reason as
above. But that introduces the problem of characters blinking out of
sync again...

Project ANISE's implementation seems to batch draw calls, so this is
probably good enough.