Add `font extract` command
Add a command to extract custom fonts (.FNT/.MSK).

This command just composes the grayscale (.FNT) and alpha (.MSK) data
into a PNG. It's not particularly useful for anything other than as a
demonstration of how the custom font format is structured (Ai Shimai
uses this format, I'm not sure about other games).